requirements.txt
Django == 1.11.7
djangorestframework == 3.7.3
mysql-connector == 2.1.6